requirements.txt
# Google Gemini AI SDK
google-genai>=0.1.0

# Data validation
pydantic>=2.0.0

# HTTP requests 
requests>=2.31.0

# Optional: Web scraping 
beautifulsoup4>=4.12.0
lxml>=4.9.0
